# config/api.yaml
openapi: 3.0.0
info:
  title: Milan API
  description:
    Feel free to raise an issue or PR
    [here](https://github.com/MilanCommunity/Milan-Backend/issues/new/choose).
  version: 1.0
servers:
  - url: http://localhost:5000/
    description: Local server
  - url: https://api.milanhub.org/
    description: Production server
components:
  securitySchemes:
    ApiTokenss:
      type: http
      scheme: bearer

    ApiKey:
      type: apiKey
      in: header
      name: apikey
paths:
...
  /display/events:
    get:
      tags:
        - Get Data API
      security:
        - ApiTokenss: []
        - ApiKey: []
      summary: Returns all the events.
      responses:
        "200":
          description: OK
...
  /event/create:
    post:
      tags:
        - Create event API
      summary: Create a new event
      description:
        <b> Request (required) </b> <br /> <br /> - <b> name </b> <br /> - <b>
        slug </b> <br /> - <b> about </b> <br /> - <b> hostClubSlug </b> <br />
        - <b> hostClubName </b> <br /> - <b> bannerImage </b> <br /> - <b>
        thumbnailImage </b> <br /> - <b> mode </b> <br /> - <b> address </b> <br
        /> - <b> city </b> <br /> - <b> state </b> <br /> - <b> country </b> <br
        /> - <b> iframe </b> <br /> - <b> data </b> <br /> - <b> time </b> <br
        /> - <b> tags </b> <br /> - <b> isApproved </b> <br /> - <b> createdAt
        </b> <br /> - <b> updatedAt </b> <br />
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Clean plastics"
                slug:
                  type: string
                  example: "cleanplastics"
                about:
                  type: string
                  example: "Plastic waste management"
                hostClubSlug:
                  type: string
                  example: "exampleclub"
                hostClubName:
                  type: string
                  example: "Example club"
                bannerImage:
                  type: string
                  example: "https://example.image.com"
                thumbnailImage:
                  type: string
                  example: "https://thumbnail.image.com"
                mode:
                  type: string
                  example: "physical mode"
                address:
                  type: string
                  example: "123 Main street"
                city:
                  type: string
                  example: "pune"
                state:
                  type: string
                  example: "Maharashtra"
                country:
                  type: string
                  example: "India"
                iframe:
                  type: string
                  example: "Example iframe"
                date:
                  type: date
                  example: "12.09.2023"
                time:
                  type: string
                  example: "10:20 AM"
                tags:
                  type: array
                  items:
                    type: string
                    example:
                      - "cleaning"
                      - "waste management"
                createdAt:
                  type: date
                  example: "12.09.2023"
                updatedAt:
                  type: date
                  eample: "13.09.2023"
      responses:
        "409":
          description: Event slug already exists
        "201":
          description: Event created
        default:
          description: Internal Server Error
  /event:
    get:
      tags:
       - Get Data API
      security : 
       - ApiTokens: []
       - ApiKey: []
      summary: Returns event details or all events
      parameters:
       - in: query
         name: slug
         required: false
         schema:
          type: string
         description : Optional. If provided, fetches a specific event by its slug.
      responses:
        '200':
          description: Success response with event details or all events.
        '404':
          description: Event Not Found.
        default:
          description: Internal Server Error.